fixes #3564
git-svn-id: https://devel.scolab.eu/gvsig-online/gvsigonline_v2/tags/gvsigol_v2.4.0@5083 216424d3-d1a8-4ca4-951f-a609719bba38

+++ gvsigol/gvsigol_services/views.py
@@ -619,7 +619,8 @@ def layer_delete(request, layer_id):
 def layer_delete_operation(request, layer_id):
     layer = Layer.objects.get(pk=layer_id)
     gs = geographic_servers.get_instance().get_server_by_id(layer.datastore.workspace.server.id)
-    gs.deleteGeoserverLayerGroup(layer.layer_group)
+    if layer.layer_group.name != '__default__':
+        gs.deleteGeoserverLayerGroup(layer.layer_group)
     gs.deleteResource(layer.datastore.workspace, layer.datastore, layer)
     gs.deleteLayerStyles(layer)
     signals.layer_deleted.send(sender=None, layer=layer)
